requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
numpy==2.1.2
python-dotenv==1.0.1
requests==2.32.3
